Factor out vespalib::small_string implementation to separate files.
