+++ vespalib/src/tests/stllike/string_test.cpp
@@ -449,4 +449,10 @@ TEST("starts_with has expected semantics for std::string_view") {
     EXPECT_FALSE(ar.starts_with("oobar"));
 }
 
+TEST("test allowed nullptr construction") {
+    EXPECT_TRUE(vespalib::string(nullptr, 0).empty());
+    EXPECT_TRUE(std::string(nullptr, 0).empty());
+    EXPECT_TRUE(std::string_view(nullptr, 0).empty());
+}
+
 TEST_MAIN() { TEST_RUN_ALL(); }
