+++ vespalib/src/vespa/vespalib/stllike/small_string.hpp
